refactor(임시저장): 파일구조 분리/클래스 캡슐화
