+++ src/main/java/utils/RaceUtils.java
@@ -1,24 +1,11 @@
+package utils;
+
+import domain.Car;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class Race {
-    public int carCount;
-    public int raceTurn;
-
-    public List<Car>generateCars(){
-        List<Car> cars = new ArrayList<>();
-        for(int i = 0; i < carCount; i++){
-            Car car = new Car("Car No." +  i, 0);
-            cars.add(car);
-        }
-        return cars;
-    }
-
-    public void initRace(List<Car> cars){
-        for(Car car : cars){
-            car.carPosition = 0;
-        }
-    }
+public class RaceUtils {
 
     public void playSingleTurn(List<Car> cars){
         for(Car car : cars){
@@ -27,7 +14,7 @@ public class Race {
         }
     }
 
-    public void getWinner(List<Car> cars){
+    public List<Car> getWinner(List<Car> cars){
         List<Car> winners = new ArrayList<>();
         int maxPosition = getMaxPosition(cars);
         for (Car car : cars) {
@@ -35,6 +22,7 @@ public class Race {
                 winners.add(car);
             }
         }
+        return winners;
     }
 
     public int getMaxPosition(List<Car> cars){
@@ -48,13 +36,9 @@ public class Race {
     public void playRace(List<Car> cars , int raceTurn){
         initRace(cars);
         for(int i=0;i<raceTurn;i++){
-            PlaySingleTurn(cars);
+            playSingleTurn(cars);
         }
 
     }
 
-    public Race(int carCount, int raceTurn) {
-        this.carCount = carCount;
-        this.raceTurn = raceTurn;
-    }
 }
