+++ src/main/java/utils/RandomDigitGenerator.java
@@ -1,3 +1,5 @@
+package utils;
+
 import java.util.Random;
 
 public class RandomDigitGenerator {
