+++ src/test/java/RaceTest.java
@@ -1,3 +1,4 @@
+import domain.Race;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
